refactor(ui): promote SqlPreviewContext, SqlGenerationType, refresh_split_button to dbflux_components (#102)
Moves SqlGenerationType and SqlPreviewContext from sql_preview_modal.rs into
dbflux_components::sql_preview (new module), re-exported from crate root.
Moves refresh_split_button into dbflux_components::composites.

Keeps pub use shims at the old paths (sql_preview_modal::{SqlGenerationType,
SqlPreviewContext} and ui::common::refresh_split_button) so existing
call-sites in dbflux_ui are unchanged. dbflux_components gains no
anyhow/dbflux_app/dbflux_storage/dbflux_mcp deps.
