+++ crates/dbflux_components/src/lib.rs
@@ -28,7 +28,11 @@ pub mod result_view;
 pub mod saved_chart;
 pub mod theme;
 
+pub mod sql_preview;
+
 pub use saved_chart::{
     SavedChart, SavedChartManager, SavedChartRefreshPolicy, SavedChartStore, TimeRangePreset,
     open_saved_charts_store,
 };
+pub use composites::refresh_split_button;
+pub use sql_preview::{SqlGenerationType, SqlPreviewContext};
